+++ spec/features/all_projects/export/esfa_users_can_export_download_spec.rb
@@ -18,35 +18,4 @@ RSpec.feature "ESFA users can export" do
 
     expect(page).to have_link("Exports")
   end
-
-  scenario "they can view six months of exports" do
-    user = create(:user, team: :education_and_skills_funding_agency)
-
-    sign_in_with_user(user)
-    visit all_all_in_progress_projects_path
-
-    click_on "Exports"
-
-    click_on "funding agreement letter contacts, RPA and start-up grants"
-
-    expect(page).to have_link(Date.today.to_fs(:govuk_month))
-    expect(page).to have_link(Date.today.next_month.to_fs(:govuk_month))
-  end
-
-  scenario "they can view six months of exports" do
-    user = create(:user, team: :education_and_skills_funding_agency)
-
-    sign_in_with_user(user)
-    visit all_all_in_progress_projects_path
-
-    click_on "Exports"
-
-    click_on "funding agreement letter contacts, RPA and start-up grants"
-
-    within("tr.govuk-table__row:contains('#{Date.today.to_fs(:govuk_month)}')") do
-      click_on "Export"
-    end
-
-    expect(page).to have_link("Download CSV file")
-  end
 end