+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - A simple POST Api endpoint to create a Form a MAT Conversion project
+- The RPA, SUG and FA letters export now includes the projects for a supplied
+  date range.
 
 ### Changed